Move the application initialization code into the SingleInitContentProvider
This will initialize modules (and later, Dagger components, guaranteed only once per lifespan of the application
due to the nature of ContentProviders. This also allows us to drop a context from being passed around for each
access to the main PYDroid module.

A couple of lines are changed in each file, but the main work is in PYDroidApplication and SingleInitContentProvider