Break the loader API up into Resource and Path loaders
Resource loaders can load from the Android R system, and currently only support Drawables.
The API is the same but the class location may have changed.

Path loaders can load from a local file path, and currently only support Bitmaps
The API is similar to the ResourceLoader, with the additional option to specify
a width and height for the Bitmap.

The empty() Loaded call has been moved into the Helper