Remove the content provider, make the application the singleton instance point
For sme reason, it seems like certain devices did not start the CP in all cases.
The application class however, was reliably loaded.

The CP also would be loaded on boot, even if the app itself was never launched
To work around this, we move initialization to the application class, and check for the correct
process via an initialization hook