Move pydroid-ui initialization into the Injector
All components of PYDroid UI rely on the Injector at some point in their initialization lifecycle
Instead of requiring users of the library to subclass from the PYDroidApplication, we can
instead initialize all of these required dependencies when the initial Injector is created.

The application context provided will bind the static singleton to the lifecycle of the process it is on.