Enforce threads on methods
The current hook setup are not thread safe, so we have two options. One
is to synchronize the hooks, which will make the insertion of hooks on
any thread, but causes sync overhead and is sort of useless, as inflate,
save, and teardown as Events only actually ever run on UI threads.

We have instead opted for the second of the two options, which is accept
that hooks are not thread safe, and can and should only be run on the
main thread. This is a breaking change and may require a version bump.

All hooks are marked with the UiThread annotation and are enforced.