Treat Main dispatcher like a single threaded queue
We enforce a strict bind-then-render ordering via the Main dispatcher

All operations that require a guaranteed order and queued up on the Main
dispatcher via queueInOrder.

To avoid excess work on the Main thread, functions which are queued in order
will almost always immediately call withContext using a different dispatcher to
do their actual work off the main thread, but since the launch was called
on Main, it will guarantee that the actions happen in their queued order

Once we have bound the controller and view event busses, we queue again
for the view initialization, queue again for the state event binding,
and then finally queue once more for the initial render.

Since queueInOrder is an extension on the CoroutineScope, it will all cancel
when the main render Job is cancelled.

This also helps us avoid spinning up a new single threaded executor pool
for each view model created