Refactor
ComposeTheme moves internal to the library
PYDroid does not enforce its own Singleton

You should use the library as if it were a Singleton,
but in theory, multiple instances can be created.

The library is only expected to behave as a Singleton though.
Any usage of multiple instances is entirely undefined.