Once an in-app update is found, show UI to prompt the user to begin the download
Once the download is complete, show other UI for the user to finalize the download
We used to show a dialog, but that was so many extra steps and clicks.

Now, we show the 'completed' UI, and once the button is selected, the
app will be killed for upgrade and then restarted by the in-app library.

In testing/dev mode, we simply close the activity to simulate what would
happen