Version check module refactor
If the download is cancelled by the user or fails for any reason, we
need to be able to handle the fail state.

In the event the user cancels, we must immediately fetch a new update
object, since calling the startUpdateFlow with the same object again
will always fail.

In the event the download fails, we show the user the error message and
prompt them to retry. If they elect to retry, we immediately grab a new
update object and then force start the update flow once it is received,
if the received object is valid.

---

This also refactors the consumer side by moving the update progress into
the version check widget, and renames the NewVersionWidget to
VersionCheckWidget.

----

There is also additional code separation that handles TEST download
cases completely isolated from the normal Play update case.